+++ http/src/main/java/io/micronaut/http/filter/FilterBodyParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.http.filter.bodyparser;
+package io.micronaut.http.filter;
 
 import io.micronaut.core.annotation.Experimental;
 import io.micronaut.core.async.annotation.SingleResult;
